Remove all redundant kconfig code
Removes all parts of the kconfig language support that has been
implemented by the language server. This slims down the typescript part
of the extension to an advanced indexer.

Propfiles are removed completely, all evaluation has been removed, and
the language support for .conf files has been left to the language
server.