Completion: Provide non-visible symbols when user starts typing
The completion list will hide non-visible symbols when the user requests
completion without any prefix text. When the user starts typing, the
non-visible symbols should be listed, so we need to inform the client
that the initial list is incomplete.

Additionally, the CONFIG_ prefix needs to take partial prefixes into
account, so that a user that types CONF, then requests completion gets
their prefix finished, instead of the old behavior, which was to slap a
CONFIG_ in front of all prefixes that didn't include the full CONFIG_.
This would lead to the symbol lookup searching for CONFIG_CONF, which
doesn't match the right items.