Zephyr: Minimize writing to settings.json
Removes all redundant writes to the workspace settings. This resolves
issues with variables being automatically expanded in the settings upon
extension activation, and prevents the extension from creating a
settings file in every workspace it's activated in.

The settings file should be treated as a way for the user to configure
the application, not as a persistent storage mechanism.