Treat IO as byte stream
The RPC header's Content-Length field designates the number of bytes in
the contents, not the number of characters. As the content is utf-8
encoded, this might not always be the same number.

Use sys.stdin.buffer and sys.stdout.buffer as default streams to ensure
that the incoming data is a raw buffer, and not a decoded string,
allowing us to pull an exact number of bytes from the stream instead of
an exact number of characters. Perform decoding manually and according
to spec.