Do not provide empty choice for snippet (#23)
* Do not provide empty choice for snippet

We have a situation where all the options for a boolean or tristate
property can be invalid and so want to make sure that in that case we
don't get an invalid list of suggestions through.

Trond suggested we can get the result with CONFIG_BT_MESH_DEBUG so I've
included the definition for that and updated the tests appropriately.

The test_change_triggering_reparse test isn't quite perfect. I've chosen
to filter out the error that is still there and assert on the rest.
There might be a better way to approach it.

* Format with yapf

* Move the check into add_choices

So that we're more robust to similar issues in the future and we clean
up the business logic a bit.

* Switch to f-strings

To avoid variable name repetition.